Fixes
Split up flower into portion that is used for packing and portion that
is used only for visual attributes in hopes that this will speed up
caching when packing. Replaced TGA saving with uncompressed version to
allow slightly larger images be saved (up to memory available for full
image). adding a sample (samples/grayred.csv) that shows how to make
large number of things in a small spreadsheet.